fix casting

+++ GameEngine/src/Public/Mesh.h
@@ -6,6 +6,7 @@
 
 #include "Shader.h"
 #include "Primitive.h"
+#include "Utils.h"
 
 #pragma endregion
 
@@ -44,7 +45,7 @@ public:
 	void Draw()
 	{
 		glBindVertexArray(VAO);
-		glDrawElements(GL_TRIANGLES, Primitive->GetNumIndices(), GL_UNSIGNED_INT, 0);
+		glDrawElements(GL_TRIANGLES, (unsigned) Primitive->GetNumIndices(), GL_UNSIGNED_INT, 0);
 	}
 
 private:
